small changes in Address and ServiceCategory entities

+++ src/main/java/com/fixmate/backend/entity/Address.java
@@ -6,6 +6,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "address")
 @Getter
@@ -20,8 +22,8 @@ public class Address {
     private String address;
     private String city;
 
-    private Double latitude;
-    private Double longitude;
+    private BigDecimal latitude;
+    private BigDecimal longitude;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id",nullable = false)
